==> 01-filter_data.ipynb <==
import numpy as np
import random
import pandas as pd

from sklearn.model_selection import train_test_split

from IPython.display import Audio, display
def allDone():
    urL = 'https://www.wavsource.com/snds_2020-10-01_3728627494378403/animals/cat_meow2.wav'
    display(Audio(url=urL, autoplay=True))
# allDone()

random.seed(666)

# motifs = np.genfromtxt('data_dev/',dtype='U')
# motifxFamMatrix = np.genfromtxt('data_dev/motifxFamMatrix.csv',delimiter=',',dtype=int)
# fams = np.genfromtxt('data_dev/fams.csv',dtype='U')

# print(len(motifs))

# X_train, X_test = train_test_split(range(len(motifs)), test_size=0.15, random_state=666)

# train_motifs = motifs[X_train]
# test_motifs = motifs[X_test]

# len(train_motifs)

# Try original data from EMBER 1.
train_motifs = np.genfromtxt('data_dev/train_motifs_unfiltered.csv',dtype='U')
test_motifs = np.genfromtxt('data_dev/test_motifs_unfiltered.csv',dtype='U')
train_motifxFamMatrix = np.genfromtxt('data_dev/train_motifxFamMatrix_unfiltered.csv',delimiter=',',dtype=int)
test_motifxFamMatrix = np.genfromtxt('data_dev/test_motifxFamMatrix_unfiltered.csv',delimiter=',',dtype=int)

print(train_motifs.shape)
print(test_motifs.shape)

train_motifs = np.genfromtxt('data_dev/train_motifs.csv',dtype='U')
print(train_motifs.shape)

train_motifs = np.genfromtxt('data_dev/new_test_motifs.csv',dtype='U')
print(train_motifs.shape)


6678-6449

## Look at CD-HIT results.

# from Bio import SeqIO

# results = '1588970924.result/1588970924.fas.db2novel.clstr.sorted'

# fasta_sequences = SeqIO.parse(open(results),'fasta')
# for i,fasta in enumerate(fasta_sequences):
#     name, sequence = fasta.id, str(fasta.seq)
# #     print(i,sequence.split())
#     if i==1000:
#         break
        
# m1 = train_motifs[2268]
# m2 = test_motifs[128]

# print(m1, m2, m1[7], m2[7])

# i1 = np.where(motifs==m1)[0][0]
# i2 = np.where(motifs==m2)[0][0]

# print(i1,i2)

# fi1 = np.where(motifxFamMatrix[i1]==1)[0]
# fi2 = np.where(motifxFamMatrix[i2]==1)[0]
# print(fi1)
# print(fi2)

# print(motifxFamMatrix[i1])
# print(motifxFamMatrix[i2])

# print(fams[fi1])
# print(fams[fi2])

## Find Hamming-near motifs between test and train sets.

### https://biology.stackexchange.com/questions/23523/hamming-distance-between-two-dna-strings
def inv_hamming_distance_COPYPASTA(s1, s2):
    #Return the Hamming distance between equal-length sequences
    if len(s1) != len(s2):
        raise ValueError("Undefined for sequences of unequal length")
    return sum(ch1 == ch2 for ch1, ch2 in zip(s1, s2)) / len(s1)

def inv_hamming_distance(s1, s2):
    #Return the Hamming distance between equal-length sequences
    if len(s1) != len(s2):
        raise ValueError("Undefined for sequences of unequal length")
    zipped = zip(s1,s2)
    ### don't want to count "empty" characters ...
    max_buff_length = max( sum([1 for c in s1 if c=='_']), sum([1 for c in s2 if c=='_'])  )
    return (sum((ch1 == ch2 and ch1 != '_' and ch2 != '_') for ch1, ch2 in zip(s1, s2)) / 
            (len(s2)-max_buff_length) )

t1 = 'GEDEESESD______'
t2 = 'AEEKEAKSD______'

print(inv_hamming_distance(t1,t2))
print(inv_hamming_distance_COPYPASTA(t1,t2))

print(inv_hamming_distance_COPYPASTA('GEDEESESD','AEEKEAKSD'))
print(inv_hamming_distance('GEDEESESD','AEEKEAKSD'))

import time
import itertools

start = time.time()
similar_motifs = []
for i,combo in enumerate(itertools.product(train_motifs, test_motifs)):
    score = inv_hamming_distance(combo[0],combo[1])
    if score >= 0.6:
#         print(score, combo)
        similar_motifs.append( [combo[0], combo[1], score] )
print("%5.3f mins" % ((time.time()-start)/60))
allDone()


sim_train = list(set([x[0] for x in similar_motifs]))
sim_test = list(set([x[1] for x in similar_motifs]))

print(len(sim_train))
print(len(sim_test))

* * * RESULTS * * *

at 0.6 thresh:
sim_train: 276
sim_test: 221

## Define and save training set indices to remove ... 

idc_to_remove = []
for motif in sim_train:
    idx = np.where(train_motifs==motif)[0][0]
    idc_to_remove.append(idx)
idc_to_remove = np.array((idc_to_remove))
print(idc_to_remove)

idc_to_keep = [x for x in range(len(train_motifs)) if x not in idc_to_remove]
X_train = idc_to_keep

print(len(X_train))

train_motifs = train_motifs[X_train]
train_motifxFamMatrix = train_motifxFamMatrix[X_train]

data_dir = "data_old/"

df = pd.DataFrame(train_motifs,dtype='U')
df.to_csv(data_dir + 'train_motifs.csv',header=None,index=None)
df = pd.DataFrame(test_motifs,dtype='U')
df.to_csv(data_dir + 'test_motifs.csv',header=None,index=None)

df = pd.DataFrame(train_motifxFamMatrix,dtype=int)
df.to_csv(data_dir + 'train_motifxFamMatrix.csv',header=None,index=None)
df = pd.DataFrame(test_motifxFamMatrix,dtype=int)
df.to_csv(data_dir + 'test_motifxFamMatrix.csv',header=None,index=None)

print(len(test_motifs))
print(len(train_motifs))

print(len(train_motifs)+len(test_motifs))

# idc_to_remove = []
# for motif in sim_train:
#     idx = np.where(motifs==motif)[0][0]
#     idc_to_remove.append(idx)
# idc_to_remove = np.array((idc_to_remove))
# print(idc_to_remove)

# idc_to_keep = [x for x in X_train if x not in idc_to_remove]
# X_train = idc_to_keep

# print(len(X_train))

# train_motifs = motifs[X_train]
# test_motifs = motifs[X_test]

# train_motifxFamMatrix = motifxFamMatrix[X_train]
# test_motifxFamMatrix = motifxFamMatrix[X_test]

# data_dir = "data_old/"

# df = pd.DataFrame(train_motifs,dtype='U')
# df.to_csv(data_dir + 'train_motifs.csv',header=None,index=None)
# df = pd.DataFrame(test_motifs,dtype='U')
# df.to_csv(data_dir + 'test_motifs.csv',header=None,index=None)

# df = pd.DataFrame(train_motifxFamMatrix,dtype=int)
# df.to_csv(data_dir + 'train_motifxFamMatrix.csv',header=None,index=None)
# df = pd.DataFrame(test_motifxFamMatrix,dtype=int)
# df.to_csv(data_dir + 'test_motifxFamMatrix.csv',header=None,index=None)

## Getting rid of motifs that I can't find on uniprot (notebook 06). (Need to be able to find all motifs so that I can elongate them for musite and deepPhos comparison.)

train_motifs = np.genfromtxt('data_dev/old_test_and_train/train_motifs.csv',dtype='U')
train_motifxFamMatrix = np.genfromtxt('data_dev/old_test_and_train/train_motifxFamMatrix.csv',delimiter=',',dtype=int)
test_motifs = np.genfromtxt('data_dev/old_test_and_train/test_motifs.csv',dtype='U')
test_motifxFamMatrix = np.genfromtxt('data_dev/old_test_and_train/test_motifxFamMatrix.csv',delimiter=',',dtype=int)

fams = np.genfromtxt('data_dev/fams.csv',dtype='U')

all_motifs = np.hstack([train_motifs,test_motifs])
all_motifxFamMatrix = np.vstack([train_motifxFamMatrix,test_motifxFamMatrix])

# bad_motifs = ['AIVHLINYQDDAELV', '_PKRKVSSAEGAAKE', 'RCCDEVMSKDPTHWA', 'GQLIGLNSPNSISRA', 
#                'HSTPASAYGSVKAYT', 'QAAYYGQTPVPGPQP', 'QSVTRATSICGSGVK', 'ESRPRAVSFRQSEPS', 
#                '__MDEDYYGSAAEWG', 'ARLKSPLSTLDWTVI' ]

bad_motifs = ['_PKRKVSSAEGAAKE','ISQVRRRSSLLGSQL','ESRPRAVSFRQSEPS','LGPGPRLSLARMPPP']

bad_idc = [ np.where(test_motifs==m)[0][0] for m in bad_motifs ]

good_idc = [x for x in range(len(test_motifs)) if x not in bad_idc]
new_test_motifs = test_motifs[good_idc]
new_test_motifxFamMatrix = test_motifxFamMatrix[good_idc]

print(len(test_motifs))
print(len(new_test_motifs))

data_dir = "data_dev/old_test_and_train/"

df = pd.DataFrame(new_test_motifs,dtype='U')
df.to_csv(data_dir + 'new_test_motifs.csv',header=None,index=None)
df = pd.DataFrame(new_test_motifxFamMatrix,dtype='U')
df.to_csv(data_dir + 'new_test_motifxFamMatrix.csv',header=None,index=None)

## Meta-data

print("train motifs:",len(train_motifs))
print("test motifs:",len(new_test_motifs))
print("total motifs:",len(new_test_motifs)+len(train_motifs))

fams = np.genfromtxt('data_dev/fams.csv',dtype='U')

train_motifs = np.genfromtxt('data_dev/train_motifs.csv',dtype='U')
train_motifxFamMatrix = np.genfromtxt('data_dev/train_motifxFamMatrix.csv',delimiter=',',dtype=int)
test_motifs = np.genfromtxt('data_dev/new_test_motifs.csv',dtype='U')
test_motifxFamMatrix = np.genfromtxt('data_dev/new_test_motifxFamMatrix.csv',delimiter=',',dtype=int)

fams_alphOrder = ['AKT','CDK','CK2',"MAPK","PIKK", "PKA","PKC","SRC"]

print("Train motifs | test motifs ... per fam")
for fam in fams_alphOrder:
    f_idx = np.where(fams==fam)[0][0]
    print(fam,sum(train_motifxFamMatrix[:,f_idx]),
         "|",sum(test_motifxFamMatrix[:,f_idx]))

Train motifs | test motifs ... per fam
AKT 379 | 64
CDK 996 | 183
CK2 718 | 138
MAPK 1248 | 218
PIKK 479 | 83
PKA 1248 | 232
PKC 1463 | 277
SRC 832 | 152

LEGACY:
Train motifs | test motifs ... per fam
AKT 382 | 64
CDK 752 | 116
CK2 775 | 98
MAPK 1275 | 187
PIKK 497 | 63
PKA 1235 | 233
PKC 1497 | 252
SRC 869 | 99
