==> 04-embedding_evaluation.ipynb <==
import numpy as np
import random
from sklearn.model_selection import train_test_split
from sklearn.metrics import silhouette_score

from IPython.display import Audio, display
def allDone():
    urL = 'https://www.wavsource.com/snds_2020-10-01_3728627494378403/animals/cat_meow2.wav'
    display(Audio(url=urL, autoplay=True))
# allDone()
random.seed(666)

train_motifs = np.genfromtxt('data_dev/train_motifs.csv',dtype='U')
train_motifxFamMatrix = np.genfromtxt('data_dev/train_motifxFamMatrix.csv',delimiter=',',dtype=int)
test_motifs = np.genfromtxt('data_dev/new_test_motifs.csv',dtype='U')
test_motifxFamMatrix = np.genfromtxt('data_dev/new_test_motifxFamMatrix.csv',delimiter=',',dtype=int)

fams = np.genfromtxt('data_dev/fams.csv',dtype='U')

from sklearn.model_selection import train_test_split

all_motifs = np.hstack([train_motifs,test_motifs])
all_motifxFamMatrix = np.vstack([train_motifxFamMatrix,test_motifxFamMatrix])

print(train_motifxFamMatrix.shape)
print(test_motifxFamMatrix.shape)      

run = '02'
embedding = np.genfromtxt('data_dev/protvec_embedding_avg_special.csv',delimiter=',',dtype=float)
# embedding = np.genfromtxt('MODELS_siam/emb_%s_embedding.csv' % run,delimiter=',',dtype=float)

train_embedding = embedding[ :len(train_motifs) ]
test_embedding = embedding[ len(train_motifs): ]

print(train_embedding.shape)
print(test_embedding.shape)

#### SILHOUETTE SCORES #####

allSilhouettes = []
for i,fam in enumerate(fams):
    silhouette = silhouette_score(test_embedding, test_motifxFamMatrix.T[i], 
                                  metric='euclidean', sample_size=None, random_state=None)
    allSilhouettes.append(silhouette)
allSilhouettes_array = np.array((allSilhouettes))

# print( np.mean(allSilouettes_array), np.median(allSilouettes_array),
#        fams,[round(x,6) for x in allSilouettes] )

print("mean silhouette:", np.mean(allSilhouettes_array).round(3))
print("median silhouette", np.median(allSilhouettes_array).round(3))

def getKnnResults(fIdx, myAverage):
    
    ## best k is num of instances ^ 1/2
    ## https://stackoverflow.com/questions/18110951/how-to-determine-k-value-for-the-k-nearest-neighbours-algorithm-for-a-matrix-in
    k = int(len(embedding)**(1/2))
    
    print(k)
    
    X_tr = train_embedding 
    X_te = test_embedding
    y_tr = train_motifxFamMatrix.T[fIdx]
    y_te = test_motifxFamMatrix.T[fIdx]
    
    kNN = KNeighborsClassifier(n_neighbors=k)
    kNN.fit(X_tr, y_tr)

    y_pred = kNN.predict(X_te)
    y_score = kNN.predict_proba(X_te)

    precision, recall, fscore, support = score(y_te,y_pred,average=myAverage)
    return (precision, recall, fscore, support, y_score)

from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import precision_recall_fscore_support as score

thisY_score = [] 
for fIdx,fam in enumerate(fams):
    precision, recall, fscore, support, myY_score = (getKnnResults(fIdx, "macro"))
    thisY_score.append(myY_score)

thisY_score = np.array((thisY_score))
thisY_score.shape

y_score = thisY_score[:,:,1].T
y_test = test_motifxFamMatrix

from sklearn.metrics import roc_curve, auc
from scipy import interp
import numpy as np

print("AUROC results:")
fpr = dict()
tpr = dict()
roc_auc = dict()
for i in range(len(fams)):
    fpr[i], tpr[i], _ = roc_curve(y_test[:, i], y_score[:, i])
    roc_auc[i] = auc(fpr[i], tpr[i])
    print(fams[i],roc_auc[i].round(3))


fpr["micro"], tpr["micro"], _ = roc_curve(y_test.ravel(), y_score.ravel())
roc_auc["micro"] = auc(fpr["micro"], tpr["micro"])
all_fpr = np.unique(np.concatenate([fpr[i] for i in range(len(fams))]))

mean_tpr = np.zeros_like(all_fpr)
for i in range(len(fams)):
    mean_tpr += interp(all_fpr, fpr[i], tpr[i])
mean_tpr /= len(fams)

fpr["macro"] = all_fpr
tpr["macro"] = mean_tpr
roc_auc["macro"] = auc(fpr["macro"], tpr["macro"])

print("macro auroc:",roc_auc["macro"].round(3),
      "\nmicro auroc:",roc_auc["micro"].round(3))

------------------------------ LEGACY ------------------------------

* * * * BEST SIAM RESULTS * * * *
AUROC results:
PKA 0.852
AKT 0.897
CDK 0.892
MAPK 0.907
SRC 0.995
CK2 0.893
PKC 0.885
PIKK 0.9
macro auroc: 0.903 
micro auroc: 0.924

* * * * Avg'd ProtVec RESULTS * * * *
AUROC results:
PKA 0.867
AKT 0.907
CDK 0.887
MAPK 0.907
SRC 0.998
CK2 0.907
PKC 0.866
PIKK 0.843
macro auroc: 0.898 
micro auroc: 0.905

* * * * ProtVec RESULTS * * * *
AUROC results:
PKA 0.816
AKT 0.832
CDK 0.857
MAPK 0.888
SRC 0.979
CK2 0.872
PKC 0.83
PIKK 0.81
macro auroc: 0.861 
micro auroc: 0.851

from itertools import chain
from sklearn.metrics import precision_recall_curve
from sklearn.metrics import average_precision_score

precision = dict()
recall = dict()
average_precision = dict()

print("AVERAGE Precision (AP) results")
for i in range(len(fams)):
    precision[i], recall[i], _ = precision_recall_curve(y_test[:, i],y_score[:, i])
    average_precision[i] = average_precision_score(y_test[:, i], y_score[:, i])
    print(fams[i],average_precision[i].round(3))


average_precision["micro"] = average_precision_score(y_test, y_score,average="micro")
average_precision["macro"] = average_precision_score(y_test, y_score, average="macro")
print("avg AP (macro):",average_precision["macro"].round(3),
      "\navg AP (micro):",average_precision["micro"].round(3))

------------------------------ LEGACY ------------------------------

* * * * BEST SIAM RESULTS * * * *
AVERAGE Precision (AP) results
PKA 0.659
AKT 0.513
CDK 0.538
MAPK 0.72
SRC 0.991
CK2 0.714
PKC 0.741
PIKK 0.663
avg AP (macro): 0.692 
avg AP (micro): 0.747

* * * * Avg'd ProtVec RESULTS * * * *
AVERAGE Precision (AP) results
PKA 0.717
AKT 0.47
CDK 0.52
MAPK 0.739
SRC 0.993
CK2 0.668
PKC 0.7
PIKK 0.576
avg AP (macro): 0.673 
avg AP (micro): 0.651

* * * * ProtVec RESULTS * * * *
AVERAGE Precision (AP) results
PKA 0.625
AKT 0.421
CDK 0.41
MAPK 0.688
SRC 0.906
CK2 0.59
PKC 0.638
PIKK 0.396
avg AP (macro): 0.584 
avg AP (micro): 0.551

# allDone()

print("mean silhouette:", np.mean(allSilhouettes_array).round(3))
print("median silhouette", np.median(allSilhouettes_array).round(3))

print("----------------\n")

print("AUROC results:")
for fam in sorted(fams):
    i = np.where(fams==fam)[0][0]
    fpr[i], tpr[i], _ = roc_curve(y_test[:, i], y_score[:, i])
    roc_auc[i] = auc(fpr[i], tpr[i])
    print(fams[i],roc_auc[i].round(3))
print("macro auroc:",roc_auc["macro"].round(3),
      "\nmicro auroc:",roc_auc["micro"].round(3))

print("\n----------------\n")

print("AVERAGE Precision (AP) results")
for fam in sorted(fams):
    i = np.where(fams==fam)[0][0]
    precision[i], recall[i], _ = precision_recall_curve(y_test[:, i],y_score[:, i])
    average_precision[i] = average_precision_score(y_test[:, i], y_score[:, i])
    print(fams[i],average_precision[i].round(3))
average_precision["micro"] = average_precision_score(y_test, y_score,average="micro")
average_precision["macro"] = average_precision_score(y_test, y_score, average="macro")
print("avg AP (macro):",average_precision["macro"].round(3),
      "\navg AP (micro):",average_precision["micro"].round(3))

allDone()
